requirements.txt
python-telegram-bot==20.0
selenium
undetected-chromedriver